Rename SampleData scope to DemoData
